requirements.txt
APScheduler==3.6.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
holidays==0.9.11
idna==2.8
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tzlocal==2.0.0
urllib3==1.25.7
